moves ops around
